backend/requirements-docker.txt
# =============================================================================
# TRAE Backend - Docker Requirements
# =============================================================================
# Optimiert für Container-Umgebung (ohne GUI wie pyautogui)
# =============================================================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0

# WebSocket Support
websockets==12.0

# Data Handling / Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Image Processing (headless)
Pillow==10.1.0
opencv-python-headless==4.6.0.66
numpy==1.24.4

# HTTP Client
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
typing-extensions==4.8.0
python-dateutil==2.8.2

# JSON Handling
orjson==3.9.10

# Logging
loguru==0.7.2

# Async Utilities
asyncio-mqtt==0.16.1

# File System Monitoring
watchdog==3.0.0

# System Monitoring
psutil==5.9.6

# Database (optional)
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Retry Logic
backoff==2.2.1

# SSH Client (optional)
asyncssh==2.14.2
paramiko==3.4.0

# WebRTC Support (optional)
aiortc>=1.6.0

# Template Engine
jinja2==3.1.2